refactor(package): split manifest validation and artifact checks
