+++ packages/rust/biors-core/src/package/summary.rs
@@ -1,0 +1,21 @@
+use super::{PackageManifest, PackageManifestSummary};
+
+/// Build a compact summary of a package manifest for inspect-style output.
+pub fn inspect_package_manifest(manifest: &PackageManifest) -> PackageManifestSummary {
+    PackageManifestSummary {
+        schema_version: manifest.schema_version,
+        name: manifest.name.clone(),
+        model_format: manifest.model.format,
+        has_model_checksum: manifest.model.checksum.is_some(),
+        tokenizer: manifest
+            .tokenizer
+            .as_ref()
+            .map(|tokenizer| tokenizer.name.clone()),
+        vocab: manifest.vocab.as_ref().map(|vocab| vocab.name.clone()),
+        runtime_backend: manifest.runtime.backend,
+        runtime_target: manifest.runtime.target,
+        preprocessing_steps: manifest.preprocessing.len(),
+        postprocessing_steps: manifest.postprocessing.len(),
+        fixtures: manifest.fixtures.len(),
+    }
+}
